- Rename IStorableService to DocumentService - Adjust namespaces
