+++ src/Documentania.Infrastructure/RegionNames.cs
@@ -7,7 +7,7 @@
 // // </summary>
 // // --------------------------------------------------------------------------------------------------------------------
 
-namespace Documentania.Contracts
+namespace Documentania.Infrastructure
 {
     public class RegionNames
     {
