+++ tst/DocumentaniaSpecFlow/DocumentSaving.cs
@@ -12,7 +12,7 @@ namespace DocumentaniaSpecFlow.Tests
     using System;
     using System.Collections.Generic;
 
-    using Documentania.Contracts;
+    using Documentania.Infrastructure.Interfaces;
     using Documentania.Infrastructure.Models;
 
     using Modules.Document;
